gigshield/requirements.txt
"""
GigShield requirements
======================
Django + DRF + Celery + Redis + PostgreSQL + Razorpay + Channels + JWT
"""
Django==4.2.11
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.3
psycopg2-binary==2.9.9
django-environ==0.11.2
razorpay==1.4.1
django-celery-beat==2.6.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-ratelimit==4.1.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
# WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
